Frontend multiview support * /calendar-views api route added * calendarView method available in CalendarDataproviderInterface
