+++ resources/CalendarDataProvider.default.php
@@ -2,11 +2,11 @@
 
 namespace App\Providers;
 
-use Wdelfuego\NovaCalendar\DataProvider\MonthCalendar;
+use Wdelfuego\NovaCalendar\DataProvider\Calendar;
 use Wdelfuego\NovaCalendar\Event;
 use App\Nova\User;
 
-class CalendarDataProvider extends MonthCalendar
+class CalendarDataProvider extends Calendar
 {
     //
     // Add the Nova resources that should be displayed on the calendar to this method
