+++ routes/api.php
@@ -20,4 +20,5 @@ use Illuminate\Support\Facades\Route;
 
 use Wdelfuego\NovaCalendar\Http\Controllers\CalendarController;
 
+Route::get('/calendar-views', [CalendarController::class, 'getCalendarViews']);
 Route::get('/calendar-data/{year?}/{month?}', [CalendarController::class, 'getMonthCalendarData']);
