+++ src/Interface/CalendarDataProviderInterface.php
@@ -28,6 +28,9 @@ interface CalendarDataProviderInterface
     
     // A multi-dimensional array with all display data for 1 week in the calendar
     public function calendarWeeks() : array;
+
+    // A 1D array with available calendar views, possible elements are: ['month']. 
+    public function calendarViews(): array;
     
     // A multi-dimensional array of event styles, see documentation
     public function eventStyles() : array;
